Move code to a generic implementation
