+++ ff/src/fields/mod.rs
@@ -9,6 +9,7 @@ use ark_serialize::{
     CanonicalSerializeWithFlags, EmptyFlags, Flags,
 };
 use ark_std::{
+    cmp::min,
     fmt::{Debug, Display},
     hash::Hash,
     ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Neg, Sub, SubAssign},
@@ -342,7 +343,29 @@ pub trait PrimeField:
 
     /// Reads bytes in big-endian, and converts them to a field element.
     /// If the bytes are larger than the modulus, it will reduce them.
-    fn from_bytes_mod_order(bytes: &[u8]) -> Self;
+    fn from_bytes_mod_order(bytes: &[u8]) -> Self {
+        let num_modulus_bytes = ((Self::Params::MODULUS_BITS + 7) / 8) as usize;
+        let num_bytes_to_directly_convert = min(num_modulus_bytes - 1, bytes.len());
+        // Copy the leading big-endian bytes directly into a field element.
+        // The number of bytes directly converted must be less than the
+        // number of bytes needed to represent the modulus, as we must begin
+        // modular reduction once the data is of the same number of bytes as the modulus.
+        let mut bytes_to_directly_convert = Vec::new();
+        bytes_to_directly_convert.extend(bytes[..num_bytes_to_directly_convert].iter().rev());
+        // Guaranteed to not be None, as the input is less than the modulus size.
+        let mut res = Self::from_random_bytes(&bytes_to_directly_convert).unwrap();
+
+        // Update the result, byte by byte.
+        // We go through existing field arithmetic, which handles the reduction.
+        // TODO: If we need higher speeds, parse more bytes at once, or implement
+        // modular multiplication by a u64
+        let window_size = Self::from(256u64);
+        for i in num_bytes_to_directly_convert..bytes.len() {
+            res *= window_size;
+            res += Self::from(bytes[i]);
+        }
+        res
+    }
 
     /// Return the QNR^t, for t defined by
     /// `2^s * t = MODULUS - 1`, and t coprime to 2.
